daemon: extract commit/sync/rollback path from daemon_apply.go
Pure code-motion split of pkg/daemon/daemon_apply.go (#5661): the
commit-and-apply pipeline, peer config push, commit-confirmed handling,
and confirmed-rollback move verbatim into daemon_apply_commit.go. The
file-bootstrap import path (bootstrapFromFile) goes here too since it is
a first-boot commit driver.

Moved declarations (byte-identical bodies, no rename, no logic change):
bootstrapFromFile, maxCommitPreflightRetries, commitWithGenBinding,
commitAndApply, applyAndSyncCommitted, applyErrSkipsPeerSync,
pushCommittedConfigToPeer, syncAndApply, deviceMapPassiveAdmissionAlarm,
commitConfirmedAndApply, commitAndApplyOperator,
commitConfirmedAndApplyOperator, executeConfirmedRollback,
resyncRolledBackConfigToPeer. Same package, no caller changes;
per-file imports pruned by goimports. Build: go build ./pkg/daemon/...
green.
